+++ pykblib/functions.py
@@ -1,12 +1,14 @@
-"""PyKBLib Utility Functions."""
+"""Defines various PyKBLib internal functions."""
 
 import json
-import subprocess
+import pipes
+import time
 
+import pexpect
 from steffentools import dict_to_ntuple
 
 
-def _api_base(service, query):
+def _api_base(service, query, attempts=0):
     """Send a query to the specified service API.
 
     Parameters
@@ -23,9 +25,17 @@ def _api_base(service, query):
         The API result in namedtuple format.
 
     """
+    if attempts > 3:
+        # Try up to three times before returning False if we have a signature
+        # error. This crops up occasionally.
+        return False
     query = json.dumps(query)
     response = _run_command(["keybase", service, "api", "-m", query])
     response = json.loads(response)
+    if hasattr(response, "error"):
+        if "bad signature" in response.error.message:
+            time.sleep(2)
+            return _api_base(service, query, attempts + 1)
     return dict_to_ntuple(response)
 
 
@@ -80,6 +90,42 @@ def _api_wallet(query):
     return _api_base("wallet", query)
 
 
+def _delete_team(team_name):
+    """Attempt to delete the specified team.
+
+    Parameters
+    ----------
+    team_name : str
+        The name of the team to be deleted.
+
+    Returns
+    -------
+    bool
+        `True` or `False`, dependent on whether the function succeeded.
+
+    """
+    # This is a special function. Where most actions are able to be
+    # executed via the _run_command function, keybase team deletion is a
+    # bit more complicated.
+
+    # Open a new pexpect process for deletion.
+    proc = pexpect.spawn("keybase team delete {}".format(team_name))
+    try:
+        proc.expect("WARNING", timeout=10)
+    except (pexpect.exceptions.TIMEOUT, pexpect.exceptions.EOF):
+        # If it didn't give us a warning, then we can't delete this team.
+        return False
+    proc.sendline("nuke {}\r\n".format(team_name))
+    output = bytes()
+    # Read output until EOF.
+    while 1:
+        try:
+            output += proc.read_nonblocking(timeout=3)
+        except (pexpect.exceptions.TIMEOUT, pexpect.exceptions.EOF):
+            break
+    return "Success!" in output.decode()
+
+
 def _get_memberships(username):
     """Get a list of the teams to which the specified user belongs.
 
@@ -126,7 +172,7 @@ def _get_username():
     return username
 
 
-def _run_command(command):
+def _run_command(command_list, attempts=0):
     """Execute a console command and retrieve the result.
 
     This function is only intended to be used with Keybase console commands. It
@@ -137,7 +183,7 @@ def _run_command(command):
 
     Parameters
     ----------
-    command : list
+    command_list : list
         The command to be executed. This command is in the form of a list, with
         the command and each argument in its own element. For example:
 
@@ -163,22 +209,11 @@ def _run_command(command):
         The command-line output.
 
     """
-    attempts = 0
-    while True:
-        try:
-            # Attempt to execute the specified command and retrieve the result.
-            return subprocess.check_output(
-                command,
-                stderr=subprocess.STDOUT,
-                shell=False,
-                timeout=10,  # Raise an exception if this takes > 10 seconds.
-            ).decode()
-        except subprocess.TimeoutExpired:
-            # When the call times out, check how many times it has failed.
-            if attempts > 3:
-                # If it's failed more than three times, raise the exception.
-                raise
-            # If it hasn't failed three times yet, restart they keybase daemon.
-            subprocess.check_output("keybase ctl restart", shell=True)
-            # Increment the number of attempts.
-            attempts += 1
+    if attempts > 3:
+        return False
+    try:
+        command = " ".join(pipes.quote(arg) for arg in command_list)
+        return pexpect.run(command).decode()
+    except pexpect.exceptions.TIMEOUT:
+        pexpect.run("keybase ctl restart")
+        return _run_command(command_list, attempts + 1)