+++ pykblib/keybase.py
@@ -1,6 +1,13 @@
-"""The Keybase class."""
+"""Contains the Keybase class definition."""
 
-from pykblib.functions import _api_team, _get_memberships, _get_username
+import string
+
+from pykblib.functions import (
+    _api_team,
+    _delete_team,
+    _get_memberships,
+    _get_username,
+)
 from pykblib.team import Team
 
 
@@ -53,6 +60,8 @@ class Keybase:
         # Create the new team.
         response = _api_team(query)
         if hasattr(response, "error"):
+            if "already exists" not in response.error.message:
+                print("Error: {}".format(response.error.message))
             return False
         # Add the team to the teams list.
         self.teams.append(team_name)
@@ -62,6 +71,34 @@ class Keybase:
         # Append the new Team to the _active_teams dictionary.
         self._active_teams[team_name] = team_instance
         return team_instance
+
+    def delete_team(self, team_name):
+        """Attempt to delete the specified team and all of its sub-teams.
+
+        Parameters
+        ----------
+        team_name : str
+            The name of the team to be deleted.
+
+        """
+        allowed_characters = string.ascii_letters + string.digits + "_-."
+        if any([letter not in allowed_characters for letter in team_name]):
+            # We want to prevent any kind of mischief with malformed names.
+            return False
+        # Compile a list of teams to be deleted.
+        teams = [team for team in self.teams if team.startswith(team_name)]
+        if team_name not in teams:
+            return False
+        # Delete each of the returned teams, starting with sub-teams.
+        for team in sorted(teams, key=len, reverse=True):
+            if not _delete_team(team):
+                return False
+            # Remove this team's record from self.teams and self._active_teams.
+            self.teams.pop(self.teams.index(team))
+            if team in self._active_teams.keys():
+                team_instance = self._active_teams.pop(team)
+                del team_instance
+        return True
 
     def team(self, team_name):
         """Return a Team class instance for the specified team.
@@ -112,7 +149,9 @@ class Keybase:
             # Update the teams list.
             for team_name in self.teams:
                 if old_name in team_name:
-                    self.teams[self.teams.index(team_name)] = team_name.replace(old_name, new_name)
+                    self.teams[
+                        self.teams.index(team_name)
+                    ] = team_name.replace(old_name, new_name)
             # Update the team name in the _active_teams dict.
             if old_name in self._active_teams.keys():
                 self._active_teams[new_name] = self._active_teams.pop(old_name)